CodeAssets/prepwithai_tutorials/requirements.txt
langchain
langchain_core
langchain_openai
langgraph>=0.3.29
langgraph-checkpoint>=2.0.0
gradio
python-dotenv
pytube
youtube_transcript_api
requests
duckduckgo_search
serpapi
tavily-python
langgraph-checkpoint-sqlite
requests
firebase_admin
azure-cosmos
aiohttp>=3.8.5
graphviz>=0.20.0
langchain-google-genai
google-generativeai
bs4
matplotlib
feedparser
playwright
scrapy